requirements.txt
gym==0.21.0
gym-retro==0.8.0
stable-baselines3==1.6.2
torch==1.13.1
torchvision==0.14.1
numpy==1.23.5
opencv-python==4.8.0.74
matplotlib==3.7.1
tensorboard==2.13.0
pillow==10.0.0
